server/requirements.txt
fastapi>=0.95.2
uvicorn>=0.22.0
httpx>=0.24.0
weaviate-client>=4.0.0
gunicorn>=20.1.0
python-dotenv>=1.0.0